Rename job.py to publisher_job.py.
Rename publish_image to process_job.
