+++ mash/services/publisher/service.py
@@ -83,7 +83,7 @@ class PublisherService(PipelineService):
         Publish image for job that matches job_id
         """
         job = self.jobs[job_id]
-        job.publish_image()
+        job.process_job()
 
     def _get_listener_msg_args(self):
         """
